+++ CHANGELOG.rst
@@ -34,4 +34,6 @@ Changelog
   robust.
 * Cleaned up authentication file handling - generate default, more error checks.
 * Moved utility function CLI arg definitions into the utility functions class.
+* Added test utility function. The functionality of this will change all the
+  time, depending on my current test needs.